prototype records correctly
prototype using I2S mic into esp32 and recording to SD card over SPI records correctly

single threaded code works correctly, but placing read/write funcs in tasks cause recording
and playback to occur at improper speeds or cause what appears to be high jitter.  not sure
if this is a firmware bug, bad programming in this repo, or something else.
single threaded, however, is fast enough.

aim for next iteration is to create EspSDWaveFile objects for both
currentReadFile and currentWriteFile, as multiple files can be accessed,
and required bandwidth is low.

next step in hardware is to wire ADC over I2S and test, develop, and debug
the PCM4220/2 classes, and eventually create a class that uses all the
inherited virtual methods to configure dishrec.