refactored most of the WAV header stuff
now in chunks, though ixml and bext are not
exclusive chunks, i plan them to be instead of
inheriting.  or at least for bextchunk.

added BWF for merging BroadcastWav headers into one.
unfinished, but does compile.

changed AudioDataPad to int_fast32_t so the compiler
would stop complaining needlessly and also because
what's the point in using least8 datatypes if the reason
for padding is speed?  so we are going with fast32.

added list of class declarations at top of some headers.