added src/Loudness for getting LUFS for metadata
even though it's kind of pointless for this device,
it's still cool.  uses libebur128 from
github.com/jiixyj/libebur128.

fixed faulty algs for converting to and from float/int
and moved them to src/IntFloatConv files.

added shitty funcs in testdata for testing.
they should be rewritten but it works so whatever.

commented out lots of esp specific code so it can be rewritten
using lower level changes.  this commit is nonfunctional for recording
without modification.