added BEXTWrapper for BWF::BroadcastWav to access protected methods in BEXTChunk
made auto originator reference formatting in BEXTChunk
defined by macro AUTO_ORIG_REF_USID, which defaults to false
since i don't have an index of country codes or organisation
codes handy.

many features will most likely end up disabled for dishrec
since they are not pertinent to it, but are included as part of
the library for the sake of completion and modularity.