corrected added bytes for even byte boundaries
was forgetting to add added byte to index which is what was
being returned, not chunk size.  because of that,
size was reporting the correct size, but ::get() for
bext chunk and ixml were both returning the bytes written
index, which did not have bytes added.  that has been corrected.

the header appears to be relatively complete