renamed AudioDatatypes.h to int_audio.h since that's all that's in there
fixed the int/float conversion stuff in AudioUtils and templated
everything with int_audio

it still requires explicit conversion for a lot of stuff but
whatever it's fine

next is to load up the ring buffer or reconfig i2s and SD write methods
to either inherit or use BWFHeader::BroadcastWav for holding metadata.
need to write a big fat chunk at the top of a new wav file and also
remember to limit header size to less than that chunk.  8k to 10k should be way
more than enough.  should limit number of sync points and tracks to not break
the limit.

at some point, wav read/load needs to be implemented but not high priority
at the moment.