added basic WavWriter.h/cpp for c++ fstream writes
needs to be written for esp32 and actually combined with
BWFHeader::BroadcastWav.  right now they are combined via
inheritance, but the demo uses mutually exclusive objects for
setting/getting metadata and file writing.

pack_data and unpack_data in AudioUtils probably need a
template specialization for int_audio to account for size of datatype
not matching the width.

need to also check and revise setting bext umid, as reading in RX7
seems not to match the ascii values:

set to:
"01234567890123456789012345678901"

reads as:
ascii values of those numbers