reworked i2s and main and added file writing methods
src/main.cpp should record properly with encapsulated,
modularized logic.

main needs work to actually properly increment scenes/filenames
and so on, but testing the logic seems to work.  have not tested
yet with live i2s input, but it doesn't crash.

mounts and unmounts SD card dynamically.
was able to get away from SD.h arduino SD library, which is
GPL'd due to v1 of SdFat library.  SdFat v2 supports ExFat and
is MIT licensed, so went with SdFat v2 for this project.

Could use sdmmc, but SPI should be fast enough for this.

every other file does not take the timestamp for some reason.

need to debounce button in hardware.

next goal is firstly to clean up redundant files that do nothing,
then to debug and do I/O and data padding/packing
in main.cpp in a readable way.  it is there, now, but needs
obvious refactoring to add a UI of some kind.

some things are cross platform even though they don't need to be,
mostly for testing and debugging on x86.