main.cpp is currently a loopback test
troubleshooting i2s r/w; will most likely change
dataingestor to configure I2S as 32-bit and copy data directly
to <int_audio> ring buffer; it is padded automatically, methinks.

i think the problem it was having with pops and shit was that
the DMA buffer was not long enough.  even though it seemed to be
the correct size, the documentation is not entirely clear as to
whether it is shared when using both TX and RX on the same I2S bus.

next up is re-enabling sd card stuff, but have to hook it up physically,
first.  wanted to get a commit in before fucking everything all up again.